requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
djangorestframework==3.9.2
gunicorn==20.0.4
idna==2.8
lxml==4.3.2
oauthlib==3.0.1
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
soupsieve==1.8
urllib3==1.24.1
whitenoise==5.0.1
